+++ tests/test_tm_utils.py
@@ -5,34 +5,43 @@ from gri_shared_library.os_tools import create_folder
 
 from conftest import DT_TEST_PARAMS_DIR, SECRETS_FILE_PATH, TEST_01_GRI_PROJECT_ID, TEST_REAL_PROJECT_C1_ID
 from decision_tree.api_utils import opentopo_pull_wrapper, download_geoparquet
-from decision_tree.process_api_results import _read_geoparquet, flatten_tm_geoparquet
+from decision_tree.process_api_results import _read_geoparquet, flatten_tm_geoparquet, TestProjectHandling
 from decision_tree.process_api_results import process_tm_results
-from decision_tree.tools import convert_to_os_path, load_secrets
+from decision_tree.tools import convert_to_os_path, load_secrets, load_yaml
+from tools import folder_cleanup
 
 params_path = os.path.join(DT_TEST_PARAMS_DIR, "params_full.yaml")
-with open(params_path, 'r') as file:
-    PARAMS = yaml.safe_load(file)
+PARAMS = load_yaml(params_path)
 SECRETS = load_secrets(SECRETS_FILE_PATH)
 
 
 def test_tm_features():
     project_ids = [TEST_01_GRI_PROJECT_ID]
+    # pre-run cleanup
+    folder_cleanup(params_path)
+
     parquet_outfile, features = _get_project_tm_features(project_ids)
 
     # Confirm that the file contains at least one polygon
     assert len(features) == 3
 
+    # post-run cleanup
+    folder_cleanup(params_path)
+
 
 def test_clean_tm_features():
-    project_ids = [TEST_01_GRI_PROJECT_ID]
+    project_ids = [TEST_01_GRI_PROJECT_ID]; test_project_handling = TestProjectHandling.ONLY
+    # pre-run cleanup
+    folder_cleanup(params_path)
+
     parquet_outfile, features = _get_project_tm_features(project_ids)
 
     outfile = PARAMS['outfile']
     project_data_dir = outfile["project_data_folder"]
     geojson_dir = convert_to_os_path(project_data_dir, outfile['geojsons'])
 
-    cleaned_features = process_tm_results(params=PARAMS, tm_geoparquet_path=parquet_outfile, geojson_dir=geojson_dir,
-                                          project_ids=project_ids, limit_to_test_projects=True)
+    cleaned_features = process_tm_results(params=PARAMS, tm_df=features, geojson_dir=geojson_dir,
+                                          project_ids=project_ids, test_project_handling=test_project_handling)
 
     assert len(cleaned_features) == 3
 
@@ -44,13 +53,19 @@ def test_clean_tm_features():
     all_exist = all(col in cleaned_features.columns for col in expected_columns)
     assert all_exist
 
+    # post-run cleanup
+    folder_cleanup(params_path)
+
 
 def test_slope_statistics(tmp_path):
-    # project_ids = [TEST_01_GRI_PROJECT_ID]; limit_to_test_projects = True
+    project_ids = [TEST_01_GRI_PROJECT_ID]; test_project_handling = TestProjectHandling.ONLY
     # REAL PROJECT BELOW - Only use for examination of an actual project
-    project_ids = [TEST_REAL_PROJECT_C1_ID]; limit_to_test_projects = False
+    # project_ids = [TEST_REAL_PROJECT_C1_ID]; test_project_handling = TestProjectHandling.EXCLUDE
     # slope_statistics, poly_results, prj_results = workflow.run_decision_tree(project_ids=project_ids)
     # REAL PROJECT ABOVE
+
+    # pre-run cleanup
+    folder_cleanup(params_path)
 
     parquet_outfile, features = _get_project_tm_features(project_ids)
 
@@ -62,8 +77,8 @@ def test_slope_statistics(tmp_path):
     project_data_dir = outfile["project_data_folder"]
     geojson_dir = convert_to_os_path(project_data_dir, outfile['geojsons'])
 
-    cleaned_features = process_tm_results(params=PARAMS, tm_geoparquet_path=parquet_outfile, geojson_dir=geojson_dir,
-                                          project_ids=project_ids, limit_to_test_projects=limit_to_test_projects)
+    cleaned_features = process_tm_results(params=PARAMS, tm_df=features, geojson_dir=geojson_dir,
+                                          project_ids=project_ids, test_project_handling=test_project_handling)
 
     # get slope statistics
     slope_statistics = opentopo_pull_wrapper(PARAMS, SECRETS, geojson_dir, cleaned_features)
@@ -77,6 +92,9 @@ def test_slope_statistics(tmp_path):
     actual_attribute_count = slope_statistics.shape[1]
     expected_attribute_count = 18
     assert actual_attribute_count == expected_attribute_count
+
+    # post-run cleanup
+    folder_cleanup(params_path)
 
 
 def _get_project_tm_features(project_ids):
@@ -95,6 +113,9 @@ def _get_project_tm_features(project_ids):
 
     # Thin to projects
     features = raw_df[raw_df['project_id'].isin(project_ids)].reset_index(drop=True)
+
+    # standardize column names
+    features = features.rename(columns={'project_name': 'short_name', "poly_id": "poly_uuid", "area": "calc_area"})
 
     return parquet_outfile, features